Refactor and propConnect Animate + AnimateGroup
