+++ src/components/Animate/utils.js
@@ -1,0 +1,1 @@
+export const COMPONENT_KEY = 'Animate'
